Adding threefold detection and tests.

+++ src/statedata.h
@@ -45,7 +45,7 @@ typedef struct {
     int whiteKingSquare;        // The current square index of the white king.
     int blackKingSquare;        // The current square index of the black king.
     int epFile;                 // The current en passant file, if any.
-    long fiftyMoveCount;        // The fifty move count. This is the number of half-moves since a capture or pawn move.
+    int fiftyMoveCount;        // The fifty move count. This is the number of half-moves since a capture or pawn move.
     int halfMoveCount;          // The half move count. This increments by one after every move.
     uint64_t hash;              // The current state hash.
     int whitePieceCount;        // The current white piece count.
